Migration

+++ src/users/users.service.ts
@@ -1,16 +1,22 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, BadRequestException } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { InjectRepository } from '@nestjs/typeorm';
-import { User } from './entities/user.entity';
+import { User } from '../../entities/user.entity';
 import { Repository } from 'typeorm';
+import { validate } from 'class-validator';
 
 @Injectable()
 export class UsersService {
   constructor(@InjectRepository(User) private usersRepository: Repository<User>,) {}
 
-  create(createUserDto: CreateUserDto) {
+  async create(createUserDto: CreateUserDto) {
     const newUser = this.usersRepository.create(createUserDto);
+
+    const errors = await validate(newUser);
+    if (errors.length > 0) {
+      throw new BadRequestException({ message: 'Validation failed', errors });
+    }
 
     return this.usersRepository.save(newUser);
   }
@@ -24,12 +30,28 @@ export class UsersService {
   }
 
   async update(id: number, updateUserDto: UpdateUserDto) {
-    const user = await this.findOne(id)
-    return this.usersRepository.save({ ...user, ...updateUserDto})
+    const user = await this.findOne(id);
+
+    if (!user) {
+      throw new BadRequestException({ message: 'User not found' });
+    }
+
+    const updatedUser = this.usersRepository.merge(user, updateUserDto);
+
+    const errors = await validate(updatedUser);
+    if (errors.length > 0) {
+      throw new BadRequestException({ message: 'Validation failed', errors });
+    }
+
+    return this.usersRepository.save(updatedUser);
   }
 
   async remove(id: number) {
-    const user = await this.findOne(id)
+    const user = await this.findOne(id);
+
+    if (!user) {
+      throw new BadRequestException({ message: 'User not found' });
+    }
 
     return this.usersRepository.remove(user);
   }
